Move reward training scripts into own package
